Remove wrong and unneeded retry for execv
execvp[e] functions cannot fail with EINTR, so we don't need to check
for this error and retry.

EAGAIN can happen if the process limit is exceeded after changing the
effective user id, and the new uder id has exceeded its process limits.
In this case, we do not want to retry the call, as this may turn into
an endless busy loop.

This patch removes the unneeded retry after execve[p] failure, making it
fail correctly with EAGAIN.

Change-Id: I5a8740afd31d6b124a8c89409b9e75c35e33649c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>