Remove unneeded and wrong retry after fork
fork cannot fail with EINTR, so we do not need to retry after EINTR.

fork may fail with EAGAIN; this can mean that a system-imposed limit on
the number of threads was encountered, or we are operating under the
SCHED_DEADLINE scheduling policy and does not have the reset-on-fork
flag set. In both cases, we want to fail and we never want to retry the
call, which can become a endless busy loop.

This patch removes the unneeded retry, so we will fail now with EAGAIN.

Change-Id: I59747b3ebd5f40071c642318288cb8e01137ca91
Signed-off-by: Nir Soffer <nsoffer@redhat.com>