Simpler and more correct tests for close_fds
We used to check that a file descriptor created in the parent process is
either inherited or closed by the child process, but we did not check if
file descriptors created by cpopen itself leak into the child process.

This patch replace the previous test with a test based on /proc/pid/fd,
ensuring that we do not leak unwanted fds.

The now unneeded helper script and subtest were removed.

Change-Id: Ibee821a36c9ab8ad10e9b62c65d1f487bbf5b3cf
Signed-off-by: Nir Soffer <nsoffer@redhat.com>