Fix close_fds test when not closing fds
The current tests works when running with Python or nosetests, but fail
when running the tests with py.test, since py.test open couple file
descriptors which are inherited by the child.

    >   self.assertEqual(child_fds, set(["0", "1", "2", str(f.fileno())]))
    E   AssertionError: Items in the first set but not the second:
    E   '3'
    E   '5'
    E   '4'
    E   '7'
    E   '6'
    E   '9'

We can fix this by forking, closing all fds, and running CPopen in the
child process, or by making a simpler test, checking that file descriptor
created in the parrent is inherited by the child. I chose the later.

Change-Id: I331380ad8183cfb0a011af15ea6efc74fdaeddc0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>