Fix incorrect fd closing
Popen accepts 4 types of values for stdin, stdout, and stderr:

- None - do not create a pipe for the child, the child inherit the
  parent stream.
- PIPE - create pipe for writing data to the child or reading data from
  the child.
- fd - pass fd to the child
- file - pass file.fileno() to the child.

When using PIPE, Popen is responsible for closing the file descriptors
it created in case of an error, or the parent side file descriptors in
the normal flow.

When using existing fd or file, Popen *must not* close the file
descriptors; they are owned by the caller and it is responsible for
closing them.  Closing the file descriptor may not only prevent the
caller from using them, but may cause the caller to close unrelated file
descriptor owned by another thread when it close its file descriptor.

Python 2.7.6 made this very easy to close the right file descriptor by
adding the to_close argument. This is a set of file descriptor that may
be closed by _execute_child. Popen.__init__ will close the file
descriptors in to_close on errors.  If _execute_child closes a file
descriptor, it must remove it from to_close, so __init__ will not try to
close a file descriptor twice, which may close unrelated file descriptor
used by another thread.

When we add support for to_close argument, we did not pay attention and
ignored it instead of using it. This may cause closing unrelated file
descriptor when running child process fails.

The current code is not compatible with using fd or file as value to
stdin, stdout, and stderr. In this case, we are closing file descriptors
we should not close.

This patch fixes both problems by normalizing error handling on all
versions:

- If to_close argument is not provided, we build it based on file
  descriptors values. When both sides of a pipe are not None we may
  close the descriptors.

- We use to_close when closing file descriptor, to close only the file
  descriptors we may close.

- When we close a file descriptor, we remove it from to_close, so
  Popen.__init__ will not try to close the file descriptor again.

This change fix also the user and fd failing tests.

Change-Id: I0a0a49ebbb9d2e9e5c251a343597d51749edb3b6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>