Add failing tests for using file or fd as stream
subprocess.Popen supports fd or file as input to stdin, stdout, or
stderr. In this case it passes the fd or file.fileno() to the child, and
does not close the fd, since they are owned by the caller.

Using file or fd, we can crate piplelines such as:

   # echo ok | cat
   p2 = CPopen(["cat"], stdin=PIPE)
   p1 = CPopen(["echo", "ok"], stdout=p2.stdin)

Unfortunately CPopen fails when using this input, because the parent
side file descriptor is closed. This patch adds failing tests
demonstrating this issue. We will merge the tests into the patch fixing
this issue later.

All tests pass when using subprocess.Popen instead of cpopen.CPopen.

Change-Id: I9a7762f10844ea75741eddf8f3fb90d8880d7714
Signed-off-by: Nir Soffer <nsoffer@redhat.com>