Do not close inherited fds from parent
When using stdin=None, stdout=None, stderr=None, the child process
should inherit the parent descriptors 0, 1, and 2. For example, when
using the following code with subprocess.Popen:

    >>> from subprocess import Popen, PIPE
    >>> p = Popen(["echo", "hello"], stdout=None, stderr=PIPE)
    >>> hello

The process has written "hello" to the parent stdout. However if we try
the same with cpopen:

    >>> from cpopen import CPopen as Popen
    >>> p = Popen(["echo", "hello"], stdout=None, stderr=PIPE)
    >>> p.communicate()
    (None, 'echo: write error: Bad file descriptor\n')
    >>> p.returncode
    1

The child process got bad file descriptor for stdout, printed an error
to stderr, and failed.

This issue is hidden by the fact that CPopen is creating pipes by
default for all descriptors. Many times these pipes are not needed since
the child process does not use them for communicating with the parent.

Trying to avoid the unneeded file descriptor, using stdin=None,
stdout=None, stderr=None, the child process will inherit bad file
descriptors, and may fail in unexpected ways.

This patch fixes this issue by removing the unwanted close for file
descriptors 0, 1, and 2.

Change-Id: Ia6f64905ebcbdaa8204a93738a80acdc39fe8163
Signed-off-by: Nir Soffer <nsoffer@redhat.com>